Ways to apply and some fixes
